gdrive-space-analyzer/pyproject.toml
[project]
name = "gdrive-space-analyzer"
version = "1.0.0"
description = "Multi-account Google Drive space analyzer for Linux"
authors = [
    {name = "Skynet Team", email = "contact@skynet.dev"}
]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.11"
keywords = ["google-drive", "storage", "analyzer", "gtk", "linux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Topic :: Desktop Environment :: Gnome",
]

dependencies = [
    "PyGObject>=3.46.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.100.0",
    "secretstorage>=3.3.3",
    "matplotlib>=3.8.0",
    "pycairo>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
gdrive-analyzer = "gdrive_analyzer.main:main"

[project.urls]
Homepage = "https://github.com/flamstyl/gdrive-space-analyzer"
Documentation = "https://github.com/flamstyl/gdrive-space-analyzer/docs"
Repository = "https://github.com/flamstyl/gdrive-space-analyzer"
Issues = "https://github.com/flamstyl/gdrive-space-analyzer/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
